+++ common/src/models/error.rs
@@ -51,6 +51,8 @@ pub enum ErrorKind {
     TooManyProjects,
     ProjectHasResources(Vec<String>),
     ProjectHasRunningDeployment,
+    ProjectHasBuildingDeployment,
+    ProjectCorrupted,
     CustomDomainNotFound,
     InvalidCustomDomain,
     CustomDomainAlreadyExists,
@@ -59,7 +61,6 @@ pub enum ErrorKind {
     NotReady,
     ServiceUnavailable,
     DeleteProjectFailed,
-    RateLimited(String),
 }
 
 impl From<ErrorKind> for ApiError {
@@ -91,14 +92,22 @@ impl From<ErrorKind> for ApiError {
                 (StatusCode::FORBIDDEN, "You cannot create more projects. Delete some projects first.")
             },
             ErrorKind::ProjectHasRunningDeployment => (
-                StatusCode::FORBIDDEN,
-                "A deployment is running. Stop it with `cargo shuttle stop` first."
+                StatusCode::INTERNAL_SERVER_ERROR,
+                "Could not automatically stop the running deployment for the project. Please reach out to Shuttle support for help."
+            ),
+            ErrorKind::ProjectHasBuildingDeployment => (
+                StatusCode::BAD_REQUEST,
+                "Project currently has a deployment that is busy building. Use `cargo shuttle deployment list` to see it and wait for it to finish"
+            ),
+            ErrorKind::ProjectCorrupted => (
+                StatusCode::BAD_REQUEST,
+                "Tried to get project into a ready state for deletion but failed. Please reach out to Shuttle support for help."
             ),
             ErrorKind::ProjectHasResources(resources) => {
                 let resources = resources.join(", ");
                 return Self {
-                    message: format!("Project has resources: {}. Use `cargo shuttle resource list` and `cargo shuttle resource delete <type>` to delete them.", resources),
-                    status_code: StatusCode::FORBIDDEN.as_u16(),
+                    message: format!("Could not automatically delete the following resources: {}. Please reach out to Shuttle support for help.", resources),
+                    status_code: StatusCode::INTERNAL_SERVER_ERROR.as_u16(),
                 }
             }
             ErrorKind::InvalidProjectName(err) => {
@@ -122,12 +131,6 @@ impl From<ErrorKind> for ApiError {
             ErrorKind::Forbidden => (StatusCode::FORBIDDEN, "Forbidden"),
             ErrorKind::NotReady => (StatusCode::INTERNAL_SERVER_ERROR, "Service not ready"),
             ErrorKind::DeleteProjectFailed => (StatusCode::INTERNAL_SERVER_ERROR, "Deleting project failed"),
-            ErrorKind::RateLimited(message) => {
-                return Self {
-                    message,
-                    status_code: StatusCode::TOO_MANY_REQUESTS.as_u16(),
-                }
-            },
         };
         Self {
             message: error_message.to_string(),