+++ proto/src/lib.rs
@@ -256,79 +256,14 @@ pub mod logger {
         type Item = LogItem;
 
         async fn receive(&mut self, items: Vec<Self::Item>) {
-            // A log vector should never be received without any items. We clone the first item
-            // here so we can use IDs to generate a rate limiting logline, which we will send to
-            // the logger as a warning to the user that they are being rate limited.
-            let Some(item) = items.first().cloned() else {
-                error!("received log vector without any items");
-
-                return;
-            };
-
             if let Err(error) = self
                 .store_logs(Request::new(StoreLogsRequest { logs: items }))
                 .await
             {
-                match error.code() {
-                    tonic::Code::Unavailable => {
-                        if error.metadata().get("x-ratelimit-limit").is_some() {
-                            let LogItem {
-                                deployment_id,
-                                log_line,
-                            } = item;
-
-                            let LogLine { service_name, .. } = log_line.unwrap();
-
-                            let timestamp = Utc::now();
-
-                            let new_item = LogItem {
-                                deployment_id,
-                                log_line: Some(LogLine {
-                                    tx_timestamp: Some(prost_types::Timestamp {
-                                        seconds: timestamp.timestamp(),
-                                        nanos: timestamp.timestamp_subsec_nanos() as i32,
-                                    }),
-                                    service_name: Backend::Runtime(service_name.clone())
-                                        .to_string(),
-                                    data: "your application is producing too many logs, log recording is being rate limited".into(),
-                                }),
-                            };
-
-                            // Give the rate limiter time to refresh, the duration we need to sleep
-                            // for here is determined by the refresh rate of the rate limiter in
-                            // the logger server. It is currently set to refresh one slot per 500ms,
-                            // so we could get away with a duration of 500ms here, but we give it
-                            // some extra time in case this rate limiting was caused by a short
-                            // burst of activity.
-                            tokio::time::sleep(Duration::from_millis(1500)).await;
-
-                            if let Err(error) = self
-                                // NOTE: the request to send this rate limiting log to the logger will
-                                // also expend a slot in the logger rate limiter.
-                                .store_logs(Request::new(StoreLogsRequest {
-                                    logs: vec![new_item],
-                                }))
-                                .await
-                            {
-                                error!(
-                                    error = &error as &dyn std::error::Error,
-                                    "failed to send rate limiting warning to logger service"
-                                );
-                            };
-                        } else {
-                            error!(
-                                error = &error as &dyn std::error::Error,
-                                "failed to send batch logs to logger"
-                            );
-                        }
-                    }
-                    _ => {
-                        error!(
-                            error = &error as &dyn std::error::Error,
-                            "failed to send batch logs to logger"
-                        );
-                    }
-                };
+                error!(
+                    error = &error as &dyn std::error::Error,
+                    "failed to send batch logs to logger"
+                );
             }
         }
     }